+++ R/plot_hexbin_gene.R
@@ -41,30 +41,8 @@
 #' library(Seurat)
 #' data("pbmc_small")
 #' pbmc_small <- make_hexbin(pbmc_small, 10, dimension_reduction = "PCA")
-#' plot_hexbin_gene(pbmc_small, type = "counts", gene = "TALDO1", action = "prop_0")
-#' # For SingleCellExperiment object
-#' \dontrun{
-#' library(TENxPBMCData)
-#' library(scater)
-#' tenx_pbmc3k <- TENxPBMCData(dataset = "pbmc3k")
-#' rm_ind <- calcAverage(tenx_pbmc3k) < 0.1
-#' tenx_pbmc3k <- tenx_pbmc3k[!rm_ind, ]
-#' colData(tenx_pbmc3k) <- cbind(
-#'   colData(tenx_pbmc3k),
-#'   perCellQCMetrics(tenx_pbmc3k)
-#' )
-#' tenx_pbmc3k <- normalize(tenx_pbmc3k)
-#' tenx_pbmc3k <- runPCA(tenx_pbmc3k)
-#' tenx_pbmc3k <- make_hexbin(tenx_pbmc3k, 20, dimension_reduction = "PCA")
-#' plot_hexbin_gene(tenx_pbmc3k,
-#'   type = "logcounts",
-#'   gene = "ENSG00000135250", action = "mean"
-#' )
-#' plot_hexbin_gene(tenx_pbmc3k,
-#'   type = "logcounts",
-#'   gene = "ENSG00000135250", action = "mode"
-#' )
-#' }
+#' plot_hexbin_gene(pbmc_small, type = "counts", gene = "TALDO1", 
+#'     action = "prop_0")
 setGeneric("plot_hexbin_gene", function (sce, 
                                          type,
                                          gene,